Add a framework for multiple writers for various test data summaries
